AI-Powered-mapping-webApp/requirements.txt
flask==2.3.2
werkzeug==2.3.7  # Critical fix for the url_decode error
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
geopandas==0.14.1
folium==0.15.1
matplotlib==3.8.2
shapely==2.0.2
pyproj==3.6.1
google-generativeai==0.3.2
numpy==1.26.4
pandas==2.2.0
gunicorn==21.2.0
mapclassify==2.6.1
rasterio==1.3.8
fiona==1.9.5